Try to cause more repetition by adding bogus font assignment
The next trick, for which I won't pull a request, would have the variable numbers (`4 6 7 9`) be put in a variable somewhere so that everything would look the same. Or even better, if everything can be +1'ed, `n=4, c.font = "62px serif", c.fillText("☁", (n++*s%1400) - 300, 220),` like four or five times with only the font size different. And perhaps moving that to a variable would help too, but I don't think that's really required.

With those changes, the string to repeat would be;

```
 serif",
                     c.fillText("☁", (n++*s%1400) - 300, 220),
                     c.font = "
```